Инкапсулировал элементы приложения, создал форму изменения продукта.
