+++ eshop/server/main/urls.py
@@ -2,6 +2,7 @@ from django.urls import path
 from .views import (
     index, about, contacts)
 
+app_name = 'main'
 
 urlpatterns = [
     path('', index, name='index'),
