+++ eshop/server/products/templates/products/catalog.html
@@ -12,7 +12,7 @@
 <div class="cat_product">
 <img src="/media/products/{{ object.image }}" alt="{{ object.alt }}" title="{{ object.title }}">
 <p>{{ object.description }}</p>
-<a href="{% url 'detail' pk=object.id %}">{{ object.name }}</a>
+<a href="{% url 'products:detail' pk=object.id %}">{{ object.name }}</a>
 </div>
     {% endfor %}
