+++ eshop/server/products/templates/products/create.html
@@ -1,0 +1,9 @@
+{% extends 'server/master.html' %}
+
+{% block title %}
+Создание продуктов
+{% endblock %}
+{% block page_content %}
+{% include 'server/form.html' with form=form %}
+
+{% endblock %}
