+++ eshop/server/products/urls/__init__.py
@@ -1,0 +1,1 @@
+from .products import *
