+++ eshop/server/products/views/__init__.py
@@ -1,0 +1,2 @@
+from .products import *
+from .categories import *
