+++ eshop/server/server/urls.py
@@ -23,6 +23,7 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('main.urls')),
     path('products/', include('products.urls')),
+    path('categories/', include('products.urls.categories')),
     path('account/', include('accounts.urls', namespace='auth'))
 ]
